+++ pp-mock/handlers/transport/v3/transport_search.go
@@ -12,7 +12,6 @@ import (
 	"buf.build/gen/go/chain4travel/camino-messenger-protocol/grpc/go/cmp/services/transport/v3/transportv3grpc"
 	transportv3 "buf.build/gen/go/chain4travel/camino-messenger-protocol/protocolbuffers/go/cmp/services/transport/v3"
 	typesv1 "buf.build/gen/go/chain4travel/camino-messenger-protocol/protocolbuffers/go/cmp/types/v1"
-	typesv2 "buf.build/gen/go/chain4travel/camino-messenger-protocol/protocolbuffers/go/cmp/types/v2"
 	typesv3 "buf.build/gen/go/chain4travel/camino-messenger-protocol/protocolbuffers/go/cmp/types/v3"
 
 	"github.com/chain4travel/camino-messenger-bot/internal/metadata"
@@ -131,18 +130,14 @@ func (*TransportSearchV3Server) TransportSearch(ctx context.Context, req *transp
 	for _, query := range req.Queries {
 		filteredTrips := mockdata.TripsExtendedV3
 		for _, queryTrip := range query.GetTrips() {
-			if queryTrip == nil {
-				continue
-			}
-			searchParametersTransport := queryTrip.GetSearchParametersTransport()
-			if searchParametersTransport == nil {
+			if queryTrip.SearchParametersTransport == nil { // its optional
 				continue
 			}
 
 			// This is just an example, not real business logic:
-			filteredTrips = filterTripsByProductCodes(filteredTrips, searchParametersTransport.GetProductCodes())
-			if searchParametersTransport.GetMaxSegments() != 0 {
-				filteredTrips = filterTripsByMaxSegments(filteredTrips, searchParametersTransport.GetMaxSegments())
+			filteredTrips = filterTripsByProductCodes(filteredTrips, queryTrip.SearchParametersTransport.ProductCodes)
+			if queryTrip.SearchParametersTransport.MaxSegments != 0 {
+				filteredTrips = filterTripsByMaxSegments(filteredTrips, queryTrip.SearchParametersTransport.MaxSegments)
 			}
 		}
 
@@ -184,29 +179,24 @@ func (*TransportSearchV3Server) TransportSearch(ctx context.Context, req *transp
 		resultIDnum++
 	}
 
-	if len(searchResults) == 0 {
-		return &transportv3.TransportSearchResponse{
-			Header: &typesv1.ResponseHeader{
-				Status: typesv1.StatusType_STATUS_TYPE_SUCCESS,
-				Alerts: []*typesv1.Alert{{
-					Message: fmt.Sprintf("No results found for search %v", req.Queries),
-					Type:    typesv1.AlertType_ALERT_TYPE_INFO,
-				}},
-			},
-		}, nil
-	}
-
-	searchID := uuid.New().String()
-
 	response := &transportv3.TransportSearchResponse{
 		Header: &typesv1.ResponseHeader{
 			Status: typesv1.StatusType_STATUS_TYPE_SUCCESS,
 		},
-		Metadata: &typesv3.SearchResponseMetadata{
-			SearchId: &typesv1.UUID{Value: searchID},
-		},
 		Results: searchResults,
 	}
+
+	if len(searchResults) == 0 {
+		response.Header.Alerts = []*typesv1.Alert{{
+			Message: fmt.Sprintf("No results found for search %v", req.Queries),
+			Type:    typesv1.AlertType_ALERT_TYPE_INFO,
+		}}
+	} else {
+		response.Metadata = &typesv3.SearchResponseMetadata{
+			SearchId: &typesv1.UUID{Value: uuid.New().String()},
+		}
+	}
+
 	log.Printf("CMAccount %s received request from CMAccount %s", md.Recipient, md.Sender)
 
 	if err := grpc.SetHeader(ctx, md.ToGrpcMD()); err != nil {
@@ -214,33 +204,4 @@ func (*TransportSearchV3Server) TransportSearch(ctx context.Context, req *transp
 	}
 
 	return response, nil
-}
-
-func filterTripsByProductCodes(trips []*transportv3.TripExtended, productCodes []*typesv2.ProductCode) []*transportv3.TripExtended {
-	if len(productCodes) == 0 {
-		return trips
-	}
-	filtered := []*transportv3.TripExtended{}
-	for _, trip := range trips {
-	segmentsLoop:
-		for _, segment := range trip.Segments {
-			for _, code := range productCodes {
-				if segment.GetInfo().GetProductCode().Code == code.GetCode() {
-					filtered = append(filtered, trip)
-					break segmentsLoop
-				}
-			}
-		}
-	}
-	return filtered
-}
-
-func filterTripsByMaxSegments(trips []*transportv3.TripExtended, maxSegments int32) []*transportv3.TripExtended {
-	filtered := []*transportv3.TripExtended{}
-	for _, trip := range trips {
-		if len(trip.Segments) <= int(maxSegments) {
-			filtered = append(filtered, trip)
-		}
-	}
-	return filtered
 }
